document_ocr_api/requirements.txt
# FastAPI and Web Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Image Processing
opencv-python-headless>=4.10.0
pillow>=11.0.0
numpy>=2.0.0

# PDF Processing
pymupdf>=1.24.0

# OCR and document parsing
paddleocr>=3.3.0
paddlex[ocr]>=3.4.0
# Install paddlepaddle separately when your platform supports it.

# Data Validation
pydantic>=2.10.0

# CORS
python-json-logger>=3.2.0
requests>=2.32.0

# Development
pytest>=8.0.0
httpx>=0.28.0
